refactor(images): share resource caption preprocessing and model calls
